[#15]; refactor: 프로젝트 구조 개선
