+++ src/main/java/it/dohyun/recipe_hub/model/MemberDto.java
@@ -1,0 +1,78 @@
+package it.dohyun.recipe_hub.model;
+
+import java.time.LocalDateTime;
+
+public class MemberDto {
+  private String id;
+  private String nickname;
+  private String avatar;
+  private String introduce;
+  private String email;
+  private boolean admin;
+  private String password;
+  private LocalDateTime created;
+
+  public String getId() {
+    return id;
+  }
+
+  public void setId(String id) {
+    this.id = id;
+  }
+
+  public String getNickname() {
+    return nickname;
+  }
+
+  public void setNickname(String nickname) {
+    this.nickname = nickname;
+  }
+
+  public String getAvatar() {
+    return avatar;
+  }
+
+  public void setAvatar(String avatar) {
+    this.avatar = avatar;
+  }
+
+  public String getIntroduce() {
+    return introduce;
+  }
+
+  public void setIntroduce(String introduce) {
+    this.introduce = introduce;
+  }
+
+  public String getEmail() {
+    return email;
+  }
+
+  public void setEmail(String email) {
+    this.email = email;
+  }
+
+  public boolean isAdmin() {
+    return admin;
+  }
+
+  public void setAdmin(boolean admin) {
+    this.admin = admin;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+  }
+
+  public LocalDateTime getCreated() {
+    return created;
+  }
+
+  public void setCreated(LocalDateTime created) {
+    this.created = created;
+  }
+}
